Move source code to src/ directory
