+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/defenseunicorns/go-oscal/cmd"
+import "github.com/defenseunicorns/go-oscal/src/cmd"
 
 func main() {
 	cmd.Execute()
